+++ src/extension.ts
@@ -1,47 +1,47 @@
-import { ExtensionContext, languages, commands } from 'vscode';
-import { CompletionItemProvider } from './features/completionItemProvider'
-import { DefinitionProvider } from './features/definitionProvider'
-import { showUrlInputBox } from './features/downloadScripts';
+import * as vscode from 'vscode';
+import { CompletionItemProvider } from './features/completionItemProvider';
+import { DefinitionProvider } from './features/definitionProvider';
 import { RenameProvider } from './features/renameProvider';
-import { ScriptsWatcher } from './features/scriptsWatcher';
 import { HoverProvider } from './features/hoverProvider';
 import { FoldingRangeProvider } from './features/foldingRangeProvider';
+import { ScriptsWatcher, onRenameFile } from './features/scriptsWatcher';
 
-export function activate(context: ExtensionContext) {
-	console.log("IW5-GSC-Syntax: Init");
+export async function activate(context: vscode.ExtensionContext) {
+    console.log("IW5-GSC-Syntax: Init");
 
-	const filesWatcher = new ScriptsWatcher();
+    const filesWatcher = new ScriptsWatcher();
 	filesWatcher.startWatching();
-	
-	const completionItemProvider = languages.registerCompletionItemProvider("gsc",
-		new CompletionItemProvider(), '.');
 
-	const definitionProvider = languages.registerDefinitionProvider("gsc",
-		new DefinitionProvider());
+    const renameFiles = vscode.workspace.onDidRenameFiles(async (event) => {      
+        await onRenameFile(event);
+    });
 
-	const renameProvider = languages.registerRenameProvider("gsc", 
-		new RenameProvider());
+    const completionItemProvider = vscode.languages.registerCompletionItemProvider("gsc",
+        new CompletionItemProvider(), '.');
 
-	const downloadCommand = commands.registerCommand('extension.downloadFiles', async () => {
-		await showUrlInputBox();
-	});
+    const definitionProvider = vscode.languages.registerDefinitionProvider("gsc",
+        new DefinitionProvider());
 
-	const foldingProvider = languages.registerFoldingRangeProvider('gsc', 
-		new FoldingRangeProvider());
+    const renameProvider = vscode.languages.registerRenameProvider("gsc", 
+        new RenameProvider());
 
-	const hoverProvider = languages.registerHoverProvider('gsc', 
-		new HoverProvider());
+    const foldingProvider = vscode.languages.registerFoldingRangeProvider('gsc', 
+        new FoldingRangeProvider());
 
-	context.subscriptions.push(
-		{dispose() { filesWatcher.stopWatching(); }},
-		completionItemProvider,
-		definitionProvider,
-		renameProvider,
-		downloadCommand,
-		hoverProvider,
-		foldingProvider);
+    const hoverProvider = vscode.languages.registerHoverProvider('gsc', 
+        new HoverProvider());
+
+    context.subscriptions.push(
+        {dispose() { filesWatcher.stopWatching(); }},
+        renameFiles,
+        completionItemProvider,
+        definitionProvider,
+        renameProvider,
+        hoverProvider,
+        foldingProvider
+    );
 }
 
 export function deactivate() {
-	console.log("IW5-GSC-Syntax: Exit");
+    console.log("IW5-GSC-Syntax: Exit");
 }