+++ src/features/completionItemProvider.ts
@@ -1,108 +1,138 @@
 import * as vscode from 'vscode';
 import * as utility from './utility';
-import * as defs from '../defs/defs'
-import * as fields from '../defs/fields'
-import { waittills } from '../defs/waittills';
-import { game } from '../defs/game';
+import * as fields from '../defs/fields';
+import { defs } from '../defs/defs';
+import { waittillCompletions } from '../defs/waittills';
+import { constantCompletions } from '../defs/gameConstant';
+import { damageCompletions } from '../defs/modDamages';
 
 let includeCompletion: vscode.CompletionItem[] = [];
 let includeCallCompletion: vscode.CompletionItem[] = [];
 
 export function addIncludeCompletion(filePath: string) {
-	const include = utility.pathToInclude(filePath.toLowerCase());
+	const include = utility.pathToInclude(filePath);
+	if (!include) return;
 	includeCompletion.push(new vscode.CompletionItem(include + ";", vscode.CompletionItemKind.File));
 	includeCallCompletion.push(new vscode.CompletionItem(include + "::", vscode.CompletionItemKind.File));
 }
 
 export function removeIncludeCompletion(filePath: string) {
-	const include = utility.pathToInclude(filePath);
+	const include = utility.pathToInclude(filePath, false);
+	if (!include) return;
 	includeCompletion = includeCompletion.filter(item => item.label !== include + ";");
 	includeCallCompletion = includeCallCompletion.filter(item => item.label !== include + "::");
 }
 
 export class CompletionItemProvider {
 	private functions: vscode.CompletionItem[];
-	private playerFields: vscode.CompletionItem[];
-	private selfFields: vscode.CompletionItem[];
-    private levelFields: vscode.CompletionItem[];
+	private playerFields: vscode.CompletionItem[] = [];
+	private selfFields: vscode.CompletionItem[] = [];
+    private levelFields: vscode.CompletionItem[] = [];
 
 	constructor() {
-		this.functions = defs.defs;
-		this.functions = this.functions.concat(defs.gsc_functions);
-		this.functions = this.functions.concat(this.createCompletionItems(game, false));
-		this.functions = this.functions.concat(this.createCompletionItems(waittills, false));
-		this.playerFields = this.createCompletionItems(fields.player);
-		this.selfFields = this.createCompletionItems(fields.self);
-        this.levelFields = this.createCompletionItems(fields.level);
+		this.functions = defs;
+		this.functions = this.functions.concat(damageCompletions);
+		this.functions = this.functions.concat(waittillCompletions);
+		this.functions = this.functions.concat(constantCompletions);
+		fields.player.forEach(items => { this.playerFields.push(new vscode.CompletionItem(items, vscode.CompletionItemKind.Field)); });
+		fields.self.forEach(items => { this.selfFields.push(new vscode.CompletionItem(items, vscode.CompletionItemKind.Field)); });
+		fields.level.forEach(items => { this.levelFields.push(new vscode.CompletionItem(items, vscode.CompletionItemKind.Field)); });
 	}
 
 	async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext): Promise<vscode.CompletionItem[]> {
 		return new Promise<vscode.CompletionItem[]>(async (resolve, reject) => {
-			const text = utility.getDocumentText(document);
-			if (utility.isCommentAtPosition(text, document, position)) {
+			try {
+				const text = document.getText();
+				if (utility.isCommentAtPosition(text, document, position)) {
+					resolve([]);
+					return;
+				}
+
+				const line = utility.getDocumentLine(document, position);
+				const linePrefix = line.substring(0, position.character).trim();
+
+				if (linePrefix.startsWith('#include ')) {
+					resolve(includeCompletion);
+					return;
+				}
+
+				if (context.triggerKind === vscode.CompletionTriggerKind.TriggerCharacter)
+				{
+					if (linePrefix.endsWith('player.')) resolve(this.playerFields);
+					else if (linePrefix.endsWith('level.')) resolve(this.levelFields);
+					else resolve(this.selfFields);
+					return;
+				}
+
+				const textBeforeCursor = line.substring(0, position.character);
+				const includeCallPattern = /.*?([a-zA-Z0-9_]+(?:\\[a-zA-Z0-9_]+)*)::/i;
+				const match = includeCallPattern.exec(textBeforeCursor);
+
+				if (match) {
+					const includeDocument = await vscode.workspace.openTextDocument(utility.includeToUri(match[1]));
+					resolve(getFunctionCompletionItems(includeDocument.getText()));
+					return;
+				}
+
+				let completions: vscode.CompletionItem[] = this.functions.concat(getFunctionCompletionItems(text, false));
+
+				const includes = utility.getIncludes(text);
+				for (const include of includes) {
+					const includeDocument = await vscode.workspace.openTextDocument(utility.includeToUri(include));
+					const includeFunctions = getFunctionCompletionItems(includeDocument.getText());
+					completions = completions.concat(includeFunctions);
+				}
+
+				const functionScope = utility.tryGetFuncScopeRanges(text, position);
+				if (functionScope) completions = completions.concat(getVariables(text, functionScope, position));	
+
+				completions = completions.concat(provideDefineCompletions(document));
+
+				resolve(completions.concat(includeCallCompletion));
+			}
+			catch (error) {
 				resolve([]);
-				return;
 			}
-
-			const line = utility.getDocumentLine(document, position);
-			const linePrefix = line.substring(0, position.character).trim();
-
-			if (linePrefix.startsWith('#include ')) {
-				resolve(includeCompletion);
-				return;
-			}
-
-			if (context.triggerKind === vscode.CompletionTriggerKind.TriggerCharacter)
-			{
-				if (linePrefix.endsWith('player.')) resolve(this.playerFields);
-				else if (linePrefix.endsWith('level.')) resolve(this.levelFields);
-				else resolve(this.selfFields);
-				return;
-			}
-
-			const completions: vscode.CompletionItem[] = this.functions.concat(getFunctionCompletionItems(text));
-
-			const textBeforeCursor = line.substring(0, position.character);
-			const includeCallPattern = /.*?([a-zA-Z0-9_]+(?:\\[a-zA-Z0-9_]+)*)::/i;
-			const match = includeCallPattern.exec(textBeforeCursor);
-
-			if (match) {
-				const includeDocument = await vscode.workspace.openTextDocument(utility.includeToUri(match[1]));
-				resolve(getFunctionCompletionItems(includeDocument.getText()));
-				return;
-			}
-
-			const functionScope = utility.tryGetFuncScopeLocation(text, position);
-			if (functionScope) completions.push(...getVariables(text, functionScope));			
-			
-			const includes = utility.getIncludes(text);
-			for (const include of includes) {
-				const includeDocument = await vscode.workspace.openTextDocument(utility.includeToUri(include));
-				completions.push(...getFunctionCompletionItems(includeDocument.getText()));
-			}
-
-			resolve(completions.concat(includeCallCompletion));
 		});
 	}
-
-	createCompletionItems(items: string[], isField: boolean = true): vscode.CompletionItem[] {
-        return items.map(items => {
-			if (isField) return new vscode.CompletionItem(items, vscode.CompletionItemKind.Field);
-			else return new vscode.CompletionItem(items, vscode.CompletionItemKind.Function);
-        });
- 	}
 }
 
-function getFunctionCompletionItems(text: string): vscode.CompletionItem[] {
+function provideDefineCompletions(document: vscode.TextDocument): vscode.CompletionItem[] {
+    const defines: vscode.CompletionItem[] = [];
+    const defineRegex = /^#define\s+(\w+)\s+(.*)$/;
+
+    for (let lineNumber = 0; lineNumber < document.lineCount; lineNumber++) {
+        const lineText = document.lineAt(lineNumber).text;
+        const match = defineRegex.exec(lineText);
+        
+        if (match) {
+            const constantName = match[1];
+            const constantValue = match[2].trim();
+            const completionItem = new vscode.CompletionItem(constantName, vscode.CompletionItemKind.Constant);
+			const markdownString = new vscode.MarkdownString();
+			markdownString.appendCodeblock(`#define ${constantName} ${constantValue}`, 'gsc');
+			completionItem.documentation = markdownString;
+			markdownString.isTrusted = true;
+            defines.push(completionItem);
+        }
+    }
+    return defines;
+}
+
+function getFunctionCompletionItems(text: string, ignorePrivate: boolean = true): vscode.CompletionItem[] {
     const lines = text.split('\n');
-    const functionPattern = new RegExp(`^(\\w+)\\s*\\(.*\\)\\s*\\{`, 'gmi');
     const completions: vscode.CompletionItem[] = [];
     let match: RegExpExecArray | null;
 
-    while ((match = functionPattern.exec(text)) !== null) {
-        const lineIndex = text.substring(0, match.index).split('\n').length - 1;
-        let docString = '';
+    while ((match = utility.FUNCTION_PATTERN.exec(text)) !== null) {
+		const funcName = match[1];
 
+		if (ignorePrivate && funcName.startsWith("_")) continue;
+
+		const completionItem = new vscode.CompletionItem(funcName, vscode.CompletionItemKind.Function);
+        const lineIndex = text.substring(0, match.index).split('\n').length - 1;
+        
+		let docString = '';
         for (let i = lineIndex - 1; i >= Math.max(0, lineIndex - 6); i--) {
             const line = lines[i].trim().toLowerCase();
             if (line.startsWith('///docstringbegin')) {
@@ -113,7 +143,6 @@ function getFunctionCompletionItems(text: string): vscode.CompletionItem[] {
             }
         }
 
-        const completionItem = new vscode.CompletionItem(match[1], vscode.CompletionItemKind.Function);
         if (docString) {
 			const markdownString = new vscode.MarkdownString();
             const docLines = docString.split('\n');
@@ -137,29 +166,47 @@ function getFunctionCompletionItems(text: string): vscode.CompletionItem[] {
     return completions;
 }
 
-function getVariables(text: string, functionScope: vscode.Location): vscode.CompletionItem[] {
-    const blockStart = functionScope.range.start.line;
-    const blockEnd = functionScope.range.end.line;
-    const blockText = text.split('\n').slice(blockStart, blockEnd + 1).join('\n');
-	const completions: vscode.CompletionItem[] = [];
+function getVariables(text: string, functionScope: utility.BasicRange, position: vscode.Position): vscode.CompletionItem[] {
+     const blockText = text.split('\n').slice(functionScope.start, functionScope.end + 1).join('\n');
+    const completions: vscode.CompletionItem[] = [];
 
-	let match;
+    let match;
 
-    // search argument
     const argPattern = /^\s*\w+\s*\(([^)]*)\)/mi;
     match = argPattern.exec(blockText);
     if (match) {
         const args = match[1].split(',').map(arg => arg.trim()).filter(arg => arg.length > 0);
-		for (const arg of args) {
-			completions.push(new vscode.CompletionItem(arg, vscode.CompletionItemKind.Variable));
-		}
+        for (const arg of args) {
+            completions.push(new vscode.CompletionItem(arg, vscode.CompletionItemKind.Variable));
+        }
     }
 
-    // search variable
     const variablePattern = /([a-zA-Z_][a-zA-Z0-9]*)\s*=\s*[^;/]*;/gmi;
     while ((match = variablePattern.exec(blockText)) !== null) {
-        completions.push(new vscode.CompletionItem(match[1], vscode.CompletionItemKind.Variable));
+        const matchIndex = match.index;
+        const matchLine = blockText.substring(0, matchIndex).split('\n').length - 1;
+        const matchCharacter = matchIndex - blockText.lastIndexOf('\n', matchIndex - 1) - 1;
+        const variablePosition = new vscode.Position(functionScope.start + matchLine, matchCharacter);
+
+        if (variablePosition.isBefore(position)) {
+            completions.push(new vscode.CompletionItem(match[1], vscode.CompletionItemKind.Variable));
+        }
     }
 
+    const waittillPattern = /waittill\s*\(\s*["'][^"']+["']\s*,\s*([^);]*)\)/gmi;
+    while ((match = waittillPattern.exec(blockText)) !== null) {
+        const waittillArgs = match[1].split(',').map(arg => arg.trim()).filter(arg => arg.length > 0);
+
+        const matchIndex = match.index;
+        const matchLine = blockText.substring(0, matchIndex).split('\n').length - 1;
+        const matchCharacter = matchIndex - blockText.lastIndexOf('\n', matchIndex - 1) - 1;
+        const waittillPosition = new vscode.Position(functionScope.start + matchLine, matchCharacter);
+
+        if (waittillPosition.isBefore(position)) {
+            for (const arg of waittillArgs) {
+                completions.push(new vscode.CompletionItem(arg, vscode.CompletionItemKind.Variable));
+            }
+        }
+    }
     return completions;
 }